refactor: rename pipefy-config → pipefy-infra
The package's role is broader than configuration loading — it owns the
adapter layer between Pipefy's application code (pipefy-sdk,
pipefy-auth) and external concerns (filesystem, OS, network boundary).
That maps to DDD/clean-architecture's "infrastructure layer".

The boundary the package defends stays honest: stdlib + pydantic /
pydantic-settings only. No httpx, no keyring, no gql, no business
logic. Anything that wants to live here must respect that contract.

Mechanical rename:
- packages/config/ → packages/infra/
- pipefy_config / pipefy-config → pipefy_infra / pipefy-infra
- pyproject.toml workspace members + sources + pytest paths
- README.md expanded to advertise the architectural role
