+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
-- **SDK / Auth**: shared filesystem-backed configuration via `~/.config/pipefy/config.toml` (`%APPDATA%\pipefy\config.toml` on Windows; honours `XDG_CONFIG_HOME` and a `PIPEFY_CONFIG_FILE=<path>` override). Top-level TOML keys map to pydantic field names on `AuthSettings` and `PipefySettings`; precedence is `init kwargs > env > .env > config.toml > defaults`. The new `pipefy-config` workspace package owns the loader (a schema-agnostic `PydanticBaseSettingsSource` subclass) and the path-discovery helpers; SDK and Auth depend on it symmetrically — neither imports the other. See [`docs/config.md`](docs/config.md) for the schema.
+- **SDK / Auth**: shared filesystem-backed configuration via `~/.config/pipefy/config.toml` (`%APPDATA%\pipefy\config.toml` on Windows; honours `XDG_CONFIG_HOME` and a `PIPEFY_CONFIG_FILE=<path>` override). Top-level TOML keys map to pydantic field names on `AuthSettings` and `PipefySettings`; precedence is `init kwargs > env > .env > config.toml > defaults`. The new `pipefy-infra` workspace package owns the loader (a schema-agnostic `PydanticBaseSettingsSource` subclass) and the path-discovery helpers; SDK and Auth depend on it symmetrically — neither imports the other. See [`docs/config.md`](docs/config.md) for the schema.
 - **CLI**: `pipefy auth status [--json|-j]` — reports auth source, identity, session expiry, and exit codes.
 - **CLI**: `pipefy auth logout` — revokes the refresh token at the IdP and clears the stored session.
 - **Auth**: `AuthSettings.auth_url` now defaults to the Pipefy production IdP when `PIPEFY_AUTH_URL` is unset — removes the previous `PIPEFY_AUTH_URL is required` exit-2 friction on `pipefy auth login` / `pipefy auth logout` and wires the MCP stored-session tier automatically after `pipefy auth login`. Override by setting `PIPEFY_AUTH_URL=<url>` to a non-prod IdP. Closes #233.
