+++ packages/infra/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "pipefy-config"
+name = "pipefy-infra"
 dynamic = ["version"]
 description = "Shared filesystem-backed configuration helpers (TOML loader, path discovery) for Pipefy SDK and Auth."
 readme = "README.md"
@@ -20,10 +20,10 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/pipefy_config"]
+packages = ["src/pipefy_infra"]
 
 [tool.hatch.version]
-path = "src/pipefy_config/__init__.py"
+path = "src/pipefy_infra/__init__.py"
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
@@ -37,7 +37,7 @@ extend = "../../pyproject.toml"
 lint.extend-select = ["TID251"]
 
 [tool.ruff.lint.flake8-tidy-imports.banned-api]
-"pipefy_auth".msg = "pipefy_config is a leaf package; it must not import the Auth package."
-"pipefy_sdk".msg = "pipefy_config is a leaf package; it must not import the SDK package."
-"pipefy_cli".msg = "pipefy_config is a leaf package; it must not import the CLI package."
-"pipefy_mcp".msg = "pipefy_config is a leaf package; it must not import the MCP package."
+"pipefy_auth".msg = "pipefy_infra is a leaf package; it must not import the Auth package."
+"pipefy_sdk".msg = "pipefy_infra is a leaf package; it must not import the SDK package."
+"pipefy_cli".msg = "pipefy_infra is a leaf package; it must not import the CLI package."
+"pipefy_mcp".msg = "pipefy_infra is a leaf package; it must not import the MCP package."
