+++ packages/sdk/pyproject.toml
@@ -9,7 +9,7 @@ dependencies = [
     "gql[httpx]>=3.5.2",
     "httpx>=0.27.0",
     "openpyxl>=3.1.5",
-    "pipefy-config",
+    "pipefy-infra",
     "pydantic>=2.11.3",
     "pydantic-settings>=2.8.1",
     "rapidfuzz>=3.14.3",
